requirements.txt
adal==1.2.7
adlfs==2022.11.2
aiobotocore==2.4.1
aiohttp==3.8.3
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
azure-core==1.26.1
azure-datalake-store==0.0.52
azure-identity==1.12.0
azure-storage-blob==12.14.1
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
crcmod==1.7
cryptography==38.0.4
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.3
dulwich==0.20.50
dvc==2.38.1
dvc-azure==2.20.5
dvc-data==0.28.4
dvc-gdrive==2.19.1
dvc-gs==2.20.0
dvc-hdfs==2.19.0
dvc-http==2.27.2
dvc-objects==0.14.0
dvc-oss==2.19.0
dvc-render==0.0.15
dvc-s3==2.21.0
dvc-ssh==2.20.0
dvc-task==0.1.8
dvc-webdav==2.19.0
dvc-webhdfs==2.19.0
dvclive==1.2.2
filelock==3.8.2
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gcsfs==2022.11.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grandalf==0.6
h11==0.14.0
httpcore==0.16.2
httplib2==0.21.0
httpx==0.23.1
huggingface-hub==0.11.1
hydra-core==1.3.0
idna==3.4
isodate==0.6.1
iterative-telemetry==0.0.6
jmespath==0.10.0
joblib==1.2.0
knack==0.10.1
kombu==5.2.4
lorem==0.1.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.3
nanotime==0.5.2
networkx==2.8.8
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
oss2==2.16.0
ossfs==2021.8.0
packaging==22.0
pandas==1.5.2
pathspec==0.9.0
portalocker==2.6.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydot==1.4.2
PyDrive2==1.15.0
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3fs==2022.11.0
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
scmrepo==0.1.4
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sshfs==2022.6.0
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomlkit==0.11.6
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
webdav4==0.9.8
wrapt==1.14.1
yarl==1.8.2
zc.lockfile==2.0
